// src/modules/characters/characters.types.ts
export interface IGetCharactersParams {
  name?: string;
  species?: string;
  gender?: string;
  status?: string;
  page?: string;
}

export interface ICharacter {
  id: number;
  name: string;
  status: string;
  species: string;
  type: string;
  gender: string;
  origin: {
    name: string;
    url: string;
  };
  location: {
    name: string;
    url: string;
  };
  image: string;
  episode: [];
  url: string;
  created: string;
}